requirements.txt
pandas>=2.2
pyarrow>=16.0
requests>=2.32
PyYAML>=6.0
duckdb>=1.0